Separar modelo de número de orden de la app persona en una app propia (nro_orden)
