+++ nro_orden/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class NroOrdenConfig(AppConfig):
+    name = 'nro_orden'
